requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_py-xgboost-mutex=2.0=cpu_0
argon2-cffi=19.1.0=pypi_0
bcrypt=3.1.6=pypi_0
blas=1.0=mkl
boto=2.49.0=pypi_0
boto3=1.9.192=pypi_0
botocore=1.12.192=pypi_0
ca-certificates=2019.10.16=0
certifi=2019.9.11=py36_0
cffi=1.12.3=pypi_0
chardet=3.0.4=pypi_0
click=7.0=py36_0
dj-database-url=0.5.0=pypi_0
django=1.11=pypi_0
django-bootstrap3=11.0.0=pypi_0
django-storages=1.7.1=pypi_0
djangorestframework=3.10.3=pypi_0
docutils=0.14=pypi_0
faker=1.0.7=pypi_0
flask=1.1.1=py_0
flask-login=0.4.1=py36_0
flask-sqlalchemy=2.4.0=py_0
flask-wtf=0.14.2=py36_1
gunicorn=19.9.0=pypi_0
idna=2.8=pypi_0
intel-openmp=2019.4=233
itsdangerous=1.1.0=py36_0
jinja2=2.10.3=py_0
jmespath=0.9.4=pypi_0
joblib=0.14.0=py_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libxgboost=0.90=h0a44026_1
llvm-openmp=4.0.1=hcfea43d_1
markupsafe=1.1.1=py36h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py36hfbe908c_0
mkl_fft=1.0.15=py36h5e564d8_0
mkl_random=1.1.0=py36ha771720_0
ncurses=6.1=h0a44026_1
numpy=1.16.0=pypi_0
numpy-base=1.17.3=py36h6575580_0
openssl=1.1.1d=h1de35cc_3
pandas=0.25.0=pypi_0
pillow=6.0.0=pypi_0
pip=19.1.1=py36_0
psycopg2=2.8.3=pypi_0
py-xgboost=0.90=py36h0a44026_1
py-xgboost-cpu=0.90=py36_1
pycparser=2.19=pypi_0
python=3.6.8=haf84260_0
python-dateutil=2.8.0=pypi_0
pytz=2019.1=pypi_0
readline=7.0=h1de35cc_5
requests=2.22.0=pypi_0
s3transfer=0.2.1=pypi_0
scikit-learn=0.21.3=py36h27c97d8_0
scipy=1.3.1=pypi_0
setuptools=41.0.1=py36_0
six=1.12.0=pypi_0
sqlalchemy=1.3.10=py36h1de35cc_0
sqlite=3.29.0=ha441bb4_0
sqlparse=0.3.0=pypi_0
stripe=2.32.1=pypi_0
text-unidecode=1.2=pypi_0
tk=8.6.8=ha441bb4_0
urllib3=1.25.3=pypi_0
werkzeug=0.16.0=py_0
wheel=0.33.4=py36_0
wtforms=2.2.1=py36_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
